fix: update display of setAchievement page
