+++ apps/client/pages/Auth/SetAchievement.vue
@@ -6,7 +6,7 @@
       成就中心
       <button
         class="btn btn-primary"
-        @click="handleShowModal"
+        @click="handleOpenDialog"
       >
         颁发成就
       </button>
@@ -14,12 +14,11 @@
 
     <div class="flex flex-wrap gap-5">
       <AchievementCard
-        v-for="a in _achievementList"
+        v-for="a in achievementList"
         :key="a.id"
         class="cursor-pointer"
         isShowCheckBox
         :achievement="a"
-        @toggleChooseAchievement="toggleChooseAchievement"
       />
     </div>
   </div>
@@ -43,7 +42,7 @@
           <div class="flex gap-2 mt-2">
             <div
               class="badge badge-ghost"
-              v-for="i in chooseAchievement"
+              v-for="i in checkedAchievement"
             >
               {{ i.name }}
             </div>
@@ -88,14 +87,24 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
+import { onMounted, ref } from "vue";
+import Message from "~/components/main/Message/useMessage";
 import AchievementCard from "~/components/user/AchievementCard.vue";
-import type { AchievementItem } from "~/composables/user/achievement";
+import { useAchievementList } from "~/composables/user/achievement";
 import FormInput from "~/pages/Auth/FormInput.vue";
 import { useAwardForm } from "./hooks/useAwardForm";
 const { handleSubmit, phone, phoneError, authentication, authenticationError } =
   useAwardForm();
-const chooseAchievement = ref<AchievementItem[]>([]);
+const {achievementList, getAchievementList,checkedAchievement,getAchievementChecked} = useAchievementList()
+
+function handleOpenDialog(){
+  getAchievementChecked()
+  if(checkedAchievement.value.length > 0){
+    handleShowModal()
+  }else{
+    Message.warning("请先选择成就", { duration: 1200 });
+  }
+}
 function useShowModal() {
   const isShow = ref(false);
 
@@ -108,50 +117,12 @@ function useShowModal() {
   return { isShow, handleShowModal, handleHideModal };
 }
 const { isShow, handleShowModal, handleHideModal } = useShowModal();
-const idList = ref([]);
-const toggleChooseAchievement = (item: AchievementItem) => {
-  chooseAchievement.value.push(item);
-};
+
 
 const handleAward = handleSubmit(async (values) => {});
-const _achievementList: AchievementItem[] = [
-  {
-    id: 1,
-    name: "成就名1",
-    avatar: "https://avatars.githubusercontent.com/u/1?v=4",
-    desc: "刷课刷得停不下来",
-  },
-  {
-    id: 2,
-    name: "成就名2",
-    avatar: "https://avatars.githubusercontent.com/u/2?v=5",
-    desc: "刷课刷得最多的人",
-  },
-  {
-    id: 3,
-    name: "成就名3",
-    avatar: "https://avatars.githubusercontent.com/u/3?v=4",
-    desc: "刷课刷得最速度的同学",
-  },
-  {
-    id: 4,
-    name: "成就名4",
-    avatar: "https://avatars.githubusercontent.com/u/4?v=4",
-    desc: "刷课刷得最慢的同学",
-  },
-  {
-    id: 5,
-    name: "成就名5",
-    avatar: "https://avatars.githubusercontent.com/u/7?v=4",
-    desc: "提了很多issue的同学",
-  },
-  {
-    id: 6,
-    name: "成就名6",
-    avatar: "https://avatars.githubusercontent.com/u/8?v=4",
-    desc: "提了很多pr的同学",
-  },
-];
+onMounted(()=>{
+  getAchievementList()
+})
 </script>
 <style scoped>
 .title {
